Implements tests for SentenceService methods; Add copyTextComponent method; Final refactor code

+++ src/main/java/by/lozovenko/informationhandler/service/SentenceService.java
@@ -3,8 +3,11 @@ package by.lozovenko.informationhandler.service;
 import by.lozovenko.informationhandler.composite.TextComponent;
 
 public interface SentenceService {
-    int countVowelLetter(TextComponent sentence);
-    int countConsonantLetter(TextComponent sentence);
+    long countVowelLetter(TextComponent sentence);
+
+    long countConsonantLetter(TextComponent sentence);
+
     int findLongestWordLength(TextComponent sentence);
+
     int findWordsCount(TextComponent sentence);
 }
